+++ swppppp/app/src/main/java/com/siwon/prj/common/base/BaseViewHolder.kt
@@ -1,4 +1,4 @@
-package com.siwon.prj.common.adapter
+package com.siwon.prj.common.base
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
