+++ swppppp/app/src/main/java/com/siwon/prj/common/remote/datasource/RemoteMovieSearchDataSource.kt
@@ -1,6 +1,6 @@
-package com.siwon.prj.datasource
+package com.siwon.prj.common.remote.datasource
 
-import com.siwon.prj.model.Movie
+import com.siwon.prj.common.model.Movie
 
 interface RemoteMovieSearchDataSource {
     fun searchMovies(query: String, success: (ArrayList<Movie>) -> Unit, fail: (Throwable) -> Unit)
