+++ swppppp/app/src/main/java/com/siwon/prj/common/remote/datasource/service/MovieSearchService.kt
@@ -1,7 +1,7 @@
-package com.siwon.prj.datasource.service
+package com.siwon.prj.common.remote.datasource.service
 
-import com.siwon.prj.model.ApiInfo
-import com.siwon.prj.model.Movies
+import com.siwon.prj.common.model.ApiInfo
+import com.siwon.prj.common.model.Movies
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
