+++ swppppp/app/src/main/java/com/siwon/prj/common/remote/repository/MovieSearchRepository.kt
@@ -1,6 +1,6 @@
 package com.siwon.prj.repository
 
-import com.siwon.prj.model.Movie
+import com.siwon.prj.common.model.Movie
 
 interface MovieSearchRepository {
     fun searchMovies(query: String, success: (ArrayList<Movie>) -> Unit, fail: (Throwable) -> Unit)
